# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Deck Me API — decking quotes, materials, customers, and bookings.
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /time/clock-on:
    post:
      operationId: clockOn
      tags: [time]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ClockOnInput" }
      responses:
        "201":
          description: Clocked on
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TimeEntry" }
  /time/clock-off:
    post:
      operationId: clockOff
      tags: [time]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ClockOffInput" }
      responses:
        "200":
          description: Clocked off
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TimeEntry" }
  /time/manual:
    post:
      operationId: addManualTimeEntry
      tags: [time]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ManualTimeEntryInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TimeEntry" }
  /time/job/{jobId}:
    get:
      operationId: listTimeEntriesForJob
      tags: [time]
      parameters:
        - in: path
          name: jobId
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/TimeEntry" }
  /time/member/{memberId}:
    get:
      operationId: listTimeEntriesForMember
      tags: [time]
      parameters:
        - in: path
          name: memberId
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/TimeEntry" }

  # ── Portfolio ───────────────────────────────────────────────────────────
...
components:
  schemas:
...
    TimeEntry:
      type: object
      required: [id, teamMemberId, jobId, manualEntry, createdAt]
      properties:
        id: { type: integer }
        teamMemberId: { type: integer }
        memberName: { type: ["string", "null"] }
        jobId: { type: integer }
        jobTitle: { type: ["string", "null"] }
        clockOn: { type: ["string", "null"], format: date-time }
        clockOff: { type: ["string", "null"], format: date-time }
        durationMinutes: { type: ["integer", "null"] }
        manualEntry: { type: boolean }
        notes: { type: ["string", "null"] }
        createdAt: { type: string, format: date-time }
    ManualTimeEntryInput:
      type: object
      required: [teamMemberId, jobId, durationMinutes]
      properties:
        teamMemberId: { type: integer }
        jobId: { type: integer }
        durationMinutes: { type: integer }
        clockOn: { type: string, format: date-time }
        clockOff: { type: string, format: date-time }
        notes: { type: string }
...
    ClockOnInput:
      type: object
      required: [teamMemberId, jobId]
      properties:
        teamMemberId: { type: integer }
        jobId: { type: integer }
        notes: { type: string }
    ClockOffInput:
      type: object
      required: [teamMemberId, jobId]
      properties:
        teamMemberId: { type: integer }
        jobId: { type: integer }
        notes: { type: string }
...